// src/types.ts
export type BBox = [number, number, number, number];
export interface Region {
  latitude: number;
  longitude: number;
  latitudeDelta: number;
  longitudeDelta: number;
}

export interface MapDimensions {
  width: number;
  height: number;
}

export interface LatLng {
  latitude: number;
  longitude: number;
}